Extracted common code into base classes to clean up inherence tree.
