+++ src/main/java/rlbot/Hivemind.java
@@ -8,11 +8,12 @@ import java.util.Set;
 public interface Hivemind {
 
     /**
-     * Hivemind logic goes here. Returns a mapping from drone indexes to ControllerStates based on the information
-     * given in the GameTickPacket.
+     * <p>Hivemind logic goes here. Returns a mapping from drone indexes to ControllerStates based on the information
+     * given in the GameTickPacket. Only return ControllerStates for the indexes included in the droneIndexes
+     * argument.</p>
      *
-     * droneIndexes might change during the game if drones join or leave. Therefore you should make sure to check
-     * whether the indexes you control are changed. For instance do:
+     * <p>droneIndexes might change during the game if drones join or leave. Therefore you should make sure to check
+     * whether the indexes you control are changed. For instance do:</p>
      * <pre>
      * {@code
      * if (oldIndexes.equals(droneIndexes)) {
