+++ src/main/java/rlbot/manager/BotManager.java
@@ -5,6 +5,7 @@ import rlbot.ControllerState;
 import rlbot.cppinterop.RLBotDll;
 import rlbot.cppinterop.RLBotInterfaceException;
 import rlbot.flat.GameTickPacket;
+import rlbot.pyinterop.BaseSocketServer;
 
 import java.io.IOException;
 import java.util.Map;
@@ -18,17 +19,10 @@ import java.util.function.Supplier;
  * This class keeps track of all the bots, runs the main logic loops, and retrieves the
  * game data on behalf of the bots.
  */
-public class BotManager implements IBotManager {
+public class BotManager extends BaseBotManager {
 
     protected final Map<Integer, BotProcess> botProcesses = new ConcurrentHashMap<>();
 
-    protected boolean keepRunning;
-
-    protected final Object dinnerBell = new Object();
-    protected GameTickPacket latestPacket;
-    protected AtomicInteger refreshRate = new AtomicInteger(60);
-
-    @Override
     public void ensureBotRegistered(final int index, final int team, final Supplier<Bot> botSupplier) {
         if (botProcesses.containsKey(index)) {
             return;
@@ -69,17 +63,6 @@ public class BotManager implements IBotManager {
         }
     }
 
-    @Override
-    public void ensureStarted() {
-        if (keepRunning) {
-            return; // Already started
-        }
-
-        keepRunning = true;
-        Thread looper = new Thread(this::doLoop);
-        looper.start();
-    }
-
     /**
      * Returns a set of every bot index that is currently registered and running in this java process.
      * Will not include indices from humans, bots in other languages, or bots in other java processes.
@@ -91,78 +74,9 @@ public class BotManager implements IBotManager {
         return botProcesses.keySet();
     }
 
-    private void doLoop() {
-        // Minimum call rate when paused.
-        final long MAX_AGENT_CALL_PERIOD = 1000 / 30;
-
-        long lastCallRealTime = System.currentTimeMillis();
-        float lastTickGameTime = 0;
-        boolean usingLockstep = false;
-
-        try {
-            usingLockstep = RLBotDll.getMatchSettings().enableLockstep();
-        } catch (final RLBotInterfaceException e) {
-            e.printStackTrace();
-        }
-
-        try {
-            latestPacket = RLBotDll.getFlatbufferPacket();
-            lastTickGameTime = latestPacket.gameInfo().secondsElapsed();
-        } catch (RLBotInterfaceException e) {
-            e.printStackTrace();
-        }
-
-        // Set the initial value so that as urgency flips up and down, the variation will be centered over
-        // the threshold (0 urgency) and have minimal likelihood of drifting away and skipping a frame
-        float frameUrgency = -0.5f / this.refreshRate.get();
-
-        while (keepRunning) {
-            // Python version: https://github.com/RLBot/RLBot/blob/master/src/main/python/rlbot/botmanager/bot_manager.py#L194-L212
-            final int refreshRate = this.refreshRate.get();
-            try {
-                // Retrieve latest packet
-                if (usingLockstep) {
-                    latestPacket = RLBotDll.getCurrentFlatbufferPacket();
-                } else {
-                    latestPacket = RLBotDll.getFlatbufferPacket();
-                }
-
-                // Run the bot only if gameInfo has updated
-                final float tickGameTime = latestPacket.gameInfo().secondsElapsed();
-                final long now = System.currentTimeMillis();
-                final boolean shouldCallWhilePaused = now - lastCallRealTime >= MAX_AGENT_CALL_PERIOD;
-
-                // Urgency increases every frame, but don't let it build up a large backlog
-                frameUrgency += tickGameTime - lastTickGameTime;
-                frameUrgency = clamp(frameUrgency, -1f / refreshRate, 1f / refreshRate);
-
-                if((tickGameTime != lastTickGameTime || shouldCallWhilePaused) && frameUrgency >= 0 || usingLockstep){
-                    lastCallRealTime = now;
-                    // Urgency decreases when a tick is processed.
-                    frameUrgency -= 1f / refreshRate;
-                    synchronized (dinnerBell) {
-                        dinnerBell.notifyAll();
-                    }
-                }
-
-                lastTickGameTime = tickGameTime;
-
-                if (usingLockstep) {
-                    try {
-                        Thread.sleep(1000 / refreshRate);
-                    } catch (InterruptedException e) {
-                        throw new RuntimeException(e);
-                    }
-                }
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        }
-    }
-
     @Override
     public void shutDown() {
-        keepRunning = false;
+        super.shutDown();
         botProcesses.clear();
     }
 
@@ -174,22 +88,5 @@ public class BotManager implements IBotManager {
             botProcesses.remove(index);
         }
         RLBotDll.setPlayerInputFlatbuffer(new EmptyControls(), index);
-    }
-
-    /**
-     * Sets the maximum amount of packets your bot will receive per second
-     */
-    @Override
-    public void setRefreshRate(int refreshRate) {
-        // Cap the refresh between 30hz and 120hz
-        this.refreshRate.set(clamp(refreshRate, 30, 120));
-    }
-
-    private static int clamp(int value, int min, int max) {
-        return Math.max(min, Math.min(value, max));
-    }
-
-    private static float clamp(float value, float min, float max) {
-        return Math.max(min, Math.min(value, max));
     }
 }
