+++ src/test/java/rlbot/JavaExample.java
@@ -1,8 +1,6 @@
 package rlbot;
 
 import rlbot.manager.BotManager;
-import rlbot.manager.HivemindManager;
-import rlbot.pyinterop.HiveSocketServer;
 import rlbot.pyinterop.SocketServer;
 
 /**
@@ -16,14 +14,10 @@ public class JavaExample {
             throw new IllegalArgumentException("You must pass in a port as an argument!");
         }
 
-//        BotManager flatBotManager = new BotManager();
-//        flatBotManager.setRefreshRate(120);
-//        int port = Integer.parseInt(args[0]);
-//        SocketServer server = new SamplePythonInterface(port, flatBotManager);
-//        server.start();
-
-        HivemindManager hive = new HivemindManager(() -> new MyHivemind());
-        HiveSocketServer server = new HiveSocketServer(port, hive);
+        BotManager flatBotManager = new BotManager();
+        flatBotManager.setRefreshRate(120);
+        int port = Integer.parseInt(args[0]);
+        SocketServer server = new SamplePythonInterface(port, flatBotManager);
         server.start();
     }
 }
